% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/canadiana-package.R
\docType{package}
\name{canadiana-package}
\alias{canadiana}
\alias{canadiana-package}
\title{R client for Canadiana}
\description{
R client for Canadiana
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

