% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cearch.R
\name{cearch}
\alias{cearch}
\title{Search Canadiana}
\usage{
cearch(query, ...)
}
\arguments{
\item{query}{Query terms}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Search Canadiana
}
\examples{
\dontrun{
out <- cearch(query="railroads")
out$meta
out$facet
head(out$docs)
}
}

