% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item.R
\name{item}
\alias{item}
\title{Get individual record}
\usage{
item(id, ...)
}
\arguments{
\item{id}{(character) Record key}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get individual record
}
\examples{
\dontrun{
item("oocihm.00001")
}
}

